requirements.txt
Flask==3.0.3
google-cloud-firestore==2.14.0
gunicorn==21.2.0
